make primaryContainer a lighter blue; use for tabs
By default, the active tab at the bottom of the screen shows in secondaryContainer color, which for us is a light yellow/orange.
For this app, it makes more sense for this to be light blue. primaryContainer is close, but it can be lightened a bit more to match the lightness of secondaryContainer.

Because we use primaryContainer for TimestampBadges, these now sort of blend into the background because they're lighter. A better idea for this is to give an small outline that is 'primary' color.
If we replace the Badge component with a View we can add this outline, and we also don't have to worry about the warnings that we got before with the Badge component.

Also added some comments describing the style, color/height overrides for BottomNavigation