add rewritten dynamicConfig.ts
This file is replacing the angular service DynamicConfig - several of the functions have been rewritten, preserving the same function while modernizing the code.

The only notable differences with the new version are that there are no longer any `saveAndNotifyConfigChanged` or `saveAndNotifyConfigReady` functions, and routing is not handled here.
These were things that relied heavily on Angular. This is no longer an Angular module - and this file should strive to be fairly agnostic of the framework.

So the routing, and the handling of what to do when the config has updated, will be handled in the components that _use_ dynamicConfig's functions by checking the result of the returned promises.