initial draft of JoinPage.tsx
This sketch of the new Join page is fairly rudimentary, but it includes the same text as the old one, and a text popup to enter an OPcode.
It uses `initByUser` from the *new* dynamicConfig.ts file, not the old dynamic_config.js file.
Qr scanning has not been implemented yet