add onboarding states to AppContext
We will derive onboarding state from the result of StartPrefs.getPendingOnboardingState - if it is null (no pending state) then we can route to the rest of the application.
If there is a string value, it specifies which screen should be shown in the OnboardingStack.

Currently, only JoinPage is implemented in OnboardingStack