inject react root into index.html
We don't have any Angular UI elements to be displayed anymore, but we still need Angular for the services.

So we'll keep the initialization logic that is happening in ngApp.js, but remove <ion-nav-bar> and <ion-nav-view> from index.html.
Instead we will have an #appRoot element, and at the end of the Angular init logic in ngApp, we'll inject the App.tsx component into this #appRoot.

At this point, we shouldn't need to angularize / embed React views into Angular views, because there _are_ no Angular views anymore!