rework dynamic config
'getConfig' in dynamicConfig.ts is a replacement for the old DynamicConfig.configReady. Once loaded, the config is cached in dynamicConfig.ts as storedConfig, so it doesn't have to be retrieved every time.

The useAppConfig is now aware of config changes and will update accordingly - it does so by watching the `configChanged` variable which is exported from dynamicConfig.ts.

So anything in React will use the useAppConfig hook. Anything not in React (the remaining Angular services) will need to call `getConfig` directly from dynamicConfig.ts.
Eventually, when we have no Angular code, we should be able to just keep the config in the App context, and use it downstream anywhere we'd like

- 'loading' has been removed from useAppConfig - instead, the loading state of the config will just be represented by it being null. This way we do not have to watch 2 variables to determine if the config has been loaded