usePermissions tripping over state
the state for platform and osver was not updating in time to set up the checklist.

Rather than rely on state, we can get the information directly from window. We know it will exist because setup is not called until window[device] exists

see discussion:https://github.com/e-mission/e-mission-docs/issues/956#issuecomment-1730259966