use only the app-wide AppStatusModal
Since we have an AppStatusModal, and we have an AppContext to share variables throughout the component tree, we don't need a separate AppStatusModal in ProfileSettings. The "App Status" action can just control the state of the AppStatusModal in the App component.

We also now don't need the extra code about `window.appStatusModalOpened`. This was to prevent two modals from being open at the same time, but since we only include the modal in one place now, we don't have to keep track of this.