fix diaryHelper tests: update to refactored fns
We wrote these tests for the functions in diaryHelper based on master, but an incoming branch `label_dashboard_profile_sept_2023` refactored diaryHelper to support base modes and display detected modes, and so the functions are slightly different and the tests need to be updated.
Fortunately, the new functions serve basically the same purpose, and for testing we can pretty much just swap out by changing the names.

The colors did need to be updated though-- so let's just export/import modeColors so we can directly compare against the colors defined in diaryHelper.