fill out SurveyPage
On SurveyPage, we look for existing surveys for the user. If any exist, we take the most recent one and ask the user if they wish to edit it.
This commit fills in all the text explaining this, unifies the button styles with the previous page, adds icons to the buttons, and adds a text field showing the date of the previous response to be edited.