rename "join" to "welcome"
To avoid the JoinPage potentially being confused with the study "join page" on the NREL website, this is being renamed to WelcomePage.
The onboarding state values are also renamed accordingly.

This way we can unambiguously refer to the "join page" as the web page, and this "welcome page" as the in-app landing page.