implement new WelcomePage design
Based on the wireframe posted in this issue, https://github.com/e-mission/e-mission-docs/issues/956#issuecomment-1734167858, the Welcome page has been styled and touched up to be more inviting and comprehendable.
Text has also been made more user-friendly.

The change in webpack.config.js is necessary to be able to load the app icon as an image.
Since the icon.png is in 'resource' directory, not in the 'www' directory, webpack did not know to treat it as an asset. By adding 'resources' to the 'include' paths, the image gets bundled correctly and can be imported in WelcomePage.tsx