allow onboarding buttons to wrap
Noticed while testing on a device with a smaller screen that these buttons can start to leak off the page if the text is too long and the screen isn't big enough.
We should allow these to wrap if necessary. If they show on one line, they will appear centered because margin: auto. If they show on separate lines, they will be left-aligned with each other, but still horizontally centered against other content because margin: auto.