preloadDemoSurveyResponse() on SaveQrPage
Once we're on the SaveQrPage (ie successfully past the ConsentPage) and permissions are sorted out, we can register the user- and immediately after that we should start preloading the previous demographics response. Since this takes some time, we don't want to wait until we're on the SurveyPage to start loading it.

We'll store a value called `preloadedResponsePromise` at the root level of SurveyPage.tsx, along with a function called preloadDemoSurveyResponse. preloadDemoSurveyResponse will be called on SaveQrPage after user registry, so that promise is already initiated by the time we need it on SurveyPage.

If for some reason it was not already initiated on SaveQrPage (like we opened the app straight to SurveyPage) it will be initiated on SurveyPage.
But either way it's only initiated once, and log statements confirm this.

-added clarifying comments