undismissable Demo Survey if no prev response
Adds an optional 'undismissable' option to EnketoModal. If this is true, then the "Dismiss" button is hidden and the user must complete the survey to close the modal.

This is the desired behavior for initial responses to the demographics survey -- taking the survey is mandatory on the first time installing the app, but updating the response on subsequent installs is mandatory.
So, the value of `undismissable` depends on whether prevSurveyResponse is truthy or not.