remove unused CSS
Note that while ionic.app.scss is not used and can be removed, we must still keep './manual_lib/ionic/css/ionic.css' in index.js for now.
In particular, there are a few things that still use the old Logger service, which uses ionicPopup to show errors, and this needs the ionic CSS to show correctly.