ensure checkList filled before show AppStatusModal
We were running into an issue where the AppStatusModal showed in situations where it shouldn't. This is because if there are no checks loaded yet, overallStatus is false- so if AppStatusModal is initiated before the checks are ready, it will see that overallStatus is false and trigger the pop up.

We will add extra criteria: for the popup to show, overallStatus must be false AND there must be at least one check in checkList