add back enketo logo img
Despite being an image, this file used to be inside the /templates folder, which was deleted in https://github.com/e-mission/e-mission-phone/pull/1032/commits/c5fbad49369141f00150ef85afe280bdd90522f0.
The image is used by EnketoModal, so we need to add it back - this time, it will be located in /img.