add storage.ts, to replace KVStore
storage.ts is replacing storage.js which had the KVStore service inside it.
storage.ts will provide a set of functions performing the same duties as KVStorage's functions did.

- `get` -> `storageGet`
- `set` -> `storageSet`
- `remove` -> `storageRemove`
- `getDirect` -> `storageGetDirect`
- `syncAllWebAndNativeValues` -> `storageSyncLocalAndNative`

storage.ts will still use the "BEMUserCache" Cordova plugin in exactly the same way that KVStore did.
However, instead of using the `angular-local-storage` package as a wrapper around localStorage, we will use it directly.
A couple functions were added ('localStorageSet` and `localStorageGet`) to facilitate using localStorage directly - localStorage requires us to stringify objects before storing, and parse them on retrieval.

Other than these substitutions, and being rewritten in modern JS with some typings, the logic is exactly the same as it was in KVStore.

--To facilitate this change, storage.js is temporarily renamed to ngStorage.js so that it doesn't conflict with the storage.ts filename. ngStorage.js will be removed soon after it is not used anymore.