register all ChartJS elements as workaround
Described in https://github.com/e-mission/e-mission-docs/issues/986#issuecomment-1738074971, we have a production-specific error where the needed ChartJS elements are not recognized as being registered.
Likely an incompatibility between ChartJS and Webpack.

As a workaround, we will register all of the `registerables`. Unfortunately, we will no longer benefit from tree-shaking with this approach.