add tests for storage.ts, plus the needed mocks
To test storage.ts we need to be able to mock the usercache cordova plugin, as well as fill in the missing window.Logger.
I created mocks for both of these which are kept in the __mocks__ folder and simply called at the start of storage.test.ts.

Because storage.ts calls getAngularService, angular-react-helper.tsx was included in the process. That file imported PaperProvider, which did not work with jest.
Because we are not angularizing components anymore, we do not actually need much of the code in angular-react-helper anyway. Let's just take it out so we can safely implicate getAngularService.

As for the storage tests themselves, they check whether the storage functions are able to store and retrieve data, and crucially they ensure that local and native storage can compensate for each other if one of them gets cleared.