fix: handle undetermined state of appstatus checks
In AppStatusModal, we listen for the value of overallStatus and if it is false, we trigger the permissions popup.

But there is an issue where the AppStatusModal sometimes shows up even if all the checks are green.
This is because it can take some time for the checks to initialize and return back their value. During this time, we want to have a sort of 'undetermined' state for overallStatus, so that we don't show the popup prematurely.

We'll do this by allowing overallStatus to be either true/false OR undefined. If there are no checks yet, overallStatus will be undefined and we won't show the popup yet. Once the checks are initialized, we'll be able to determine the status and it will return true or false.

For the individual checks themselves, they also need to be able to report an 'undetermined' state. If their statusStates are false by default, and are false before they have actually been determined, then overallStatus will also be false while the check is still ongoing.
So we should remove the initial 'false' condition from the few checks that had it, allowing the initial to be 'undefined', and when we iterate through the checks, we will explicitly check if `statusState === false` instead of just checking if `!statusState`.

Result of this is that AppStatusModal only shows if any permissions are broken.

Tested by loading the app, with all of the correct permissions, 20 times both before and after the change.
Before the change, 2 of the attempts resulted in the AppStatusModal incorrectly appearing.
After the change, the AppStatusModal did not appear on any of the 20 attempts. After this, I disallowed the Location permission and the modal did still appear as expected.