fix getAppVersion in clientstats
This is used in ProfileSettings is for the "App Version" row at the very bottom of the Profile page.
getAppVersion() was not working correctly - the correct function to call is `getVersionNumber` which returns a promise. In a .then block, we can memoize this value in the local let appVersion.

Since it is memoized, we don't need a dedicated state value in ProfileSettings for appVersion. We can just call getAppVersion() directly where it's used in the SettingRow.