make getAppVersion work async
getAppVersion() calls cordova-plugin-app-version's getVersionNumber, which is asynchronous. So getAppVersion should be asynchronous too. I changed it to return a promise.
We handle this in clientStats by adopting async/await syntax.
We handle this in ProfileSettings by storing it as a ref once the promise is resolved. Then we just access it as appVersion.current.