add tests for clientStats.ts
Tests the methods exported from clientStats.ts, including getAppVersion, addStatReading, addStatEvent, and addStatError.

The tests record these stats and then query the usercache for the messages, expecting to see a new entry filled in with the fake data.

putMessage and getAllMessages needed to be added to the mockBEMUserCache, since the client stats are recorded as 'messages' rather than 'key-value' pairs