rewrite CommHelper service into commHelper.ts
CommHelper from js/services.js is being rewritten into commHelper.ts.
This commit moves over the functions from CommHelper that we need. So now the following functions are exported:
- fetchUrlCached (was already in commHelper.ts)
- getRawEntries
- getPipelineRangeTs
- getPipelineCompleteTs
- getMetrics
- getAggregateData
- registerUser
- updateUser
- getUser

The following functions were not moved over to the new commHelper.ts because they are not currently used in the codebase and I don't anticipate them being used in the foreseeable future:
- putOne
- getTimelineForDay
- habiticaRegister
- habiticaProxy
- moment2Localdate
- moment2Timestamp
- getRawEntriesForLocalDate