create StudySummary, refactor
StudySummary is now its own component separate from PrivacyPolicy. It uses the text that was previously in the first section of PrivacyPolicy. And PrivacyPolicy is now just what used to be the second section.
These components share the function getTemplateText - the function is located in StudySummary, since the summary comes first, and imported in PrivacyPolicy to be used there too.

getTemplateText was also defined on ConsentPage, but it doesn't need to be there as it's not handled there.