add 'summary' onboarding state; SummaryPage
We split off the summary to be separate from the privacy policy, so the summary will now show on a separate page preceding the privacy/consent page.
We will use a boolean flag in onboardingHelper, summaryDone, which is false initially, so after welcome is done, `getPendingOnboardingState` returns 'summary'.
Once the "Proceed" button is clicked on the SummaryPage, summaryDone is true and `getPendingOnboardingState` will return 'consent'.