fix permissions popup on 'summary' page; refactor
https://github.com/e-mission/e-mission-docs/issues/998
Adding the 'summary' page caused a regression in which the permissions popup could appear before the consent page, while the onboarding route is 'summary'.
This is because when it was added, I forgot to add it to the conditional statement in App.tsx, line 79.

Semantically we want the routes to be named, but also ordered. So a cleaner way to handle these routes is to declare them as enum values instead of strings.

Behind the scense, TypeScript enums automatically get assigned an integer value (WELCOME=0, SUMMARY=1, etc). This way, to see if a route is past consent or not, we can use `<` and `>` comparisons instead of specifically listing out and checking against different string values.

So, this commit switches over to using `enum` for the onboarding state route and uses `>` in the conditional to ensure that the AppStatusModal can only show if we are past the consent page.

Also tidied up all the code comments around the onboarding states and routes.