OnboardingState=DONE if done; null if undetermined
We had this issue (https://github.com/e-mission/e-mission-docs/issues/999) where if the app was exited in the middle of onboarding, and then opened again, this error would show.
I found out that this occured because the LabelTab was being rendered for a split second until the onboarding state was still being resolved.

It takes some time to determine the onboarding state, and it requires waiting for native calls, which is why `getPendingOnboardingState` in onboardingHelper returns a Promise.
But until this promise is resolved, we don't really know if we should show a) onboarding flow or b) the main app with tab navigation.

We had been keeping `pendingOnboardingState` as `null` if the onboarding was finished. In this case, there was no 'pending' onboarding state and we would just continue to the main app.

But a safer thing to do here is have onboarding state be null **only** if it hasn't been determined yet. If it *has* been determined, then we will explictly mark it with a route of DONE. So if it's DONE we show the main app, if it's something other than DONE, we show the onboarding flow, and if it's null, we can show a big loading spinner while we determine what state we are in.

In doing this, we no longer just keep track of *pending* onboarding states - we must always keep track of an onboarding state, whether it is DONE or not. So while making this adjustment, I renamed the variable throughout as simply 'onboardingState'.