use babel-jest instead of ts-jest
Some of the core React and React Native code is written in Flow, and the official, public packages have untranspiled Flow code. Bad!
Jest can't understand this when it tries to test our React code. So unfortunately we have to transpile it ourselves before running tests, using Babel with a couple different presets and plugins

- add babel.config.js
- change Jest config from json -> js
- add packages to support the Babel transpilation