Fixed issue with UnixInteger conversion
See comments on [the PR Draft](https://github.com/e-mission/e-mission-phone/pull/1052).
Luxon and Moment.js seem to handle the `endOf('day')` slightly differently.
The remedy for this was to simulate the old Unix Integer conversion for the
end date, but use the contemporary one for the start date.  This adjustment
fixed the download, and worked for the multiple timelines tested.