i18n: fix merge translations if structure differs
If the structure is different between lang and fallbackLang such that a key is an object in fallbackLang and a string in lang, we experience an error when we attempt to assign fill in the property to the string.
This change will skip filling in if the key is not an object in *both* lang and fallbackLang, thus preventing the error.