Set up types folder, added new tests
As discussed in PR #1061, Jiji is working on setting up a folder in
`www/js/` to contain many of our type interfaces and templates.  For
now, I've added my own `fileIOTypes.ts`.  We discussed the types added
so far, and found we were working with overlapping types! As such,
I plan to copy over her version of `DiaryTypes.ts` found in commit
ffcc871, and will use that to replace the current dataObj interface
in `controlHelper.test.ts`.

Changes made:
- Set up `www/js/types`, moved fsWindow interface there
- Added basic tests for createWriteFile