refactor filters: prep to remove buttons services
Preparing to remove the buttons services (https://github.com/e-mission/e-mission-phone/pull/1086), the filter functions need to be tweaked. `user_input` will no longer be stored on the trip object, it will be accessible in the LabelTabContext, not from these functions, so it will be passed in as a second param to these functions as userInputForTrip.

Also, we'll remove the surveyOpt variable from LabelTabContext since (i) it can just be read directly from appConfig and (ii) we don't want to keep more variables than necessary in LabelTabContext to keep it from getting too cluttered.

- the INVALID_EBIKE filter (along with invalidCheck) was removed as it is not used
- logic for toLabelCheck was written more concisely
- 'width' is no longer needed on the configuredFilters - this was for the old UI