remove populateCompositeTrips; update types
The last property on the trip/place objects that differs between phone and server is 'getNextEntry' called from the input matching function.
We can instead pass 'nextEntry' as a second parameter to the input matching functions.

This allows us to significantly simplify our typings and ensure that 'CompositeTrip' and 'ConfirmedPlace' here will be the same as they are on the server. Hooray!

-- While adjusting inputMatcher, I also renamed 'getUserInputForTrip' to 'getUserInputForTimelineEntry' (i) to be more consistent and (ii) in case we support a 'place user input' in the future