TripCard: don't show footer if no notes
On trips with no notes/additions, an empty <View> was being rendered (causing a 10px gap due to padding)
If there were no notes, `timelineNotesMap[trip._id.$oid]` would be `undefined`, which does not equal 0.
We should only show notes if it is defined AND the length is not 0, so we can just say `timelineNotesMap[trip._id.$oid]?.length`.