fix diaryHelper.test.ts
Since 8712380a29af1d748fd4e7bb284353b7393a7e98, the getDetectedModes function no longer goes through the sections of a trip and sums up distances to get percentages. We now use cleaned_section_summary and inferred_section_summary which are computed on the server for us. Updating the test to reflect this causes it to pass again.