remove use of `rootScope` in getAggregateData
This function used the Angular `rootScope` to get 2 of the server conn setting: connectUrl and aggregateAuth.

These are both derived from dynamic config file in the 'server' section. So we can useAppConfig() in MetricsTab and pass 'appConfig.server' in as a parameter 'serverConnConfig'.

Also,
- cleaned up some syntax of getAggregateData to make it more readable
- added type definitions for the server conn config

One thing to note is that I don't think any active study or program uses "no_auth". But it still looks like it is an option so I am including it.