comply to Typescript "strictNullChecks": true
This commit enables strictNullChecks for greater type safety and then fixes the dozens of instances where TS now warns us about values that are potentially null https://www.typescriptlang.org/tsconfig#strictNullChecks

This means adding types on a lot of places where we had not added types before. Typescript is more powerful when it's on a stricter configuration, but this means we have to be more explicit.

In a few of the files, I also cleaned up formatting and modernized syntax while I was making changes there.