remove french and italian
We previously had maintainers for the translations in these languages, but they have not been updated in a very long time. Because the app has changed so much since then, the old translations are almost entirely obsolete.
At this point, I think it is better to remove them from usage. If we ever get someone to update the translations we can add them back.
We'll still keep the old translations up on the e-mission-translate repo although they won't be used here.