cleanup and fix types in ProfileSettings
This file was a JSX file until very recently and so a lot of type errors and warnings went unnoticed. There were a lot, so I just went over the whole file - fixing typings, modernizing syntax, and removing things that are unused.
As part of this I expanded the typings for AppConfig.