fix reload on every tab switch
Each tab gets wrapped by an error boundary. We were doing this in the body of the App component which means every time the App component was updated/rerendered, the tabs were recomputed too.
Doing the error boundary wrapping outside of the App component ensures that it only happens once and we use the same tab components each time