fix timelineHelper test
- made enketoHelper's filterByNameAndVersion accept appConfig as a parameter instead of retrieving it itself. This is 1) simpler, 2) doesn't have to be async anymore and 3) easier to test by swapping out configs

Then, fixed the timelineHelper test by changing the mock config to have "TripConfirmSurvey" and the mock survey response to have the name "TripConfirmSurvey". Now it doesn't get filtered out and the test passes.

Note that eventually we don't want this to be hardcoded, but it will require changes to the structure of the config and maybe a backwards-compat