fix backwards compat on retrieved configs
The fix is in getConfig(), lines 257 and 265 - _backwardsCompatSurveyFill needed to be applied there.

But while I was here, something else I noticed is that these functions mutate the config object parameter, which means they are not pure functions. For best practice, I changed them so they return a new object with the change rather than mutating their inputs.

While I was doing that, I applied the AppConfig type definition we have declared in types/appConfigTypes.ts.
And I cleaned up the log statements in readConfigFromServer()