only use fakeConfig if passed to mockBEMUserCache
The dynamic config tests will want to actually test how the config is stored, so we don't want this to always fallback to some default 'fakeConfig'. Before the config gets stored we'll want to be able to tell if it's missing or not.

We'll remove that fallback, and any test that *does* want to use that 'fakeConfig' can pass it as an argument to the mockBEMUserCache() function