Change accessibilityRole for ExpandMenu "developer zone" to spinbutton instead of button
React Native Paper does not handle AccessibiltyRole, so we have to just rename the accessibility text to include "spinbutton" to identify this object as a spinbutton. I am going with the expectation that accessibility elements in other languages (e.g. Spanish, Lao) use the same names for objects, such as "spinbutton", "button", etc. because this won't be translated